chore: refactor tests into one function
