+++ crates/tests/operations/manager.rs
@@ -24,7 +24,7 @@ pub const DEFAULT_L2_NETWORK_URL_NO_FB: &str = "http://localhost:8128";
 /// Default Flashblocks WebSocket URL for testing
 pub const DEFAULT_SEQ_FLASHBLOCKS_WS_URL: &str = "ws://localhost:11111";
 /// Default Flashblocks WebSocket URL for testing
-pub const DEFAULT_FLASHBLOCKS_WS_URL: &str = "ws://localhost:11112";
+pub const DEFAULT_SEQ_2_FLASHBLOCKS_WS_URL: &str = "ws://localhost:11112";
 /// Default Flashblocks WebSocket URL for RPC node which obtains flashblocks from the sequencer
 pub const DEFAULT_RPC_FLASHBLOCKS_WS_URL: &str = "ws://localhost:11113";
 /// Default WebSocket URL for testing
